+++ src/components/Datatable/utils/formatToDatatableParams.ts
@@ -1,5 +1,6 @@
 import type { DataTableRequest } from '@/@types/request/datatable'
-import type { MUIDataTableColumnState, MUIDataTableState } from 'mui-datatables'
+import type { DataTableState } from 'mui-datatable-delight'
+import type { MUIDataTableColumnState } from 'mui-datatables'
 
 /**
  *
@@ -9,10 +10,10 @@ import type { MUIDataTableColumnState, MUIDataTableState } from 'mui-datatables'
  * @see https://datatables.net/manual/server-side
  */
 export default function formatToDatatableParams(
-    tableState: MUIDataTableState,
+    tableState: DataTableState,
 ): DataTableRequest {
     const orderIndex = tableState.columns.findIndex(
-        column => column.name === tableState.sortOrder.name,
+        column => column.name === tableState.sortOrder?.name,
     )
 
     return {
@@ -26,10 +27,10 @@ export default function formatToDatatableParams(
         order: [
             {
                 column: orderIndex,
-                dir: tableState.sortOrder.direction || 'desc',
+                dir: tableState.sortOrder?.direction || 'desc',
             },
         ],
-        columns: formatColumns(tableState.columns),
+        columns: formatColumns(tableState.columns as MUIDataTableColumnState[]),
     }
 }
