+++ src/components/UnitTxs/UnitTxs.tsx
@@ -1,5 +1,5 @@
 // types
-import type { MUIDataTableColumn } from 'mui-datatables'
+import type { DataTableColumnObject } from 'mui-datatable-delight'
 import type { GetRowDataType, MutateType } from '@/components/Datatable'
 import type Transaction from '@/dataTypes/Transaction'
 // vendors
@@ -139,7 +139,7 @@ export default function UnitTxs({
     )
 }
 
-const DATATABLE_COLUMNS: MUIDataTableColumn[] = [
+const DATATABLE_COLUMNS: DataTableColumnObject[] = [
     {
         name: 'uuid',
         label: 'UUID',
