+++ src/components/pages/farm-input-product-in-outs/Datatable.tsx
@@ -1,5 +1,4 @@
 // types
-import type { MUIDataTableColumn } from 'mui-datatables'
 import type ProductMovement from '@/dataTypes/ProductMovement'
 // vendors
 import { memo, useState } from 'react'
@@ -19,6 +18,7 @@ import formatNumber from '@/utils/formatNumber'
 // enums
 import { ApiUrlEnum } from './Datatable.type'
 import { ProductMovementTypeEnum } from '@/dataTypes/ProductMovement'
+import { DataTableColumnObject } from 'mui-datatable-delight'
 
 let getRowData: GetRowDataType<ProductMovement>
 
@@ -142,7 +142,7 @@ const detailsProductStateCustomBodyRenderLite = (dataIndex: number) => {
     )
 }
 
-export const DATATABLE_COLUMNS: MUIDataTableColumn[] = [
+export const DATATABLE_COLUMNS: DataTableColumnObject[] = [
     {
         name: 'uuid',
         label: 'UUID',
