+++ src/components/pages/user-loans/installments/Datatable.tsx
@@ -69,6 +69,7 @@ const UserLoanInstallmentDatatable = memo(
                             }
                         }
                     }}
+                    //@ts-expect-error fix this later
                     columns={DATATABLE_COLUMNS}
                     defaultSortOrder={DEFAULT_SORT_ORDER}
                 />
