+++ src/pages/farm-input-product-sales.tsx
@@ -125,6 +125,7 @@ export default function FarmInputProductSales() {
                 tableId="farm-input-product-sale-table"
                 apiUrl="/farm-inputs/product-sales/datatable"
                 onRowClick={handleRowClick}
+                // @ts-expect-error fix this later
                 columns={DATATABLE_COLUMNS}
                 defaultSortOrder={{ name: 'at', direction: 'desc' }}
                 getRowDataCallback={fn => (getRowData = fn)}
