+++ src/pages/farm-inputs/products.tsx
@@ -52,6 +52,7 @@ export default function FarmInputsProducts() {
         <AuthLayout title="Produk">
             <Datatable
                 apiUrl={DatatableEndpointEnum.PRODUCTS}
+                // @ts-expect-error fix this later
                 columns={columns}
                 defaultSortOrder={{ name: 'name', direction: 'asc' }}
                 onRowClick={(_, { dataIndex }, event) => {
