+++ src/pages/finances/debts.tsx
@@ -66,6 +66,7 @@ export default function Debts() {
                     })
                     setDebtData(data)
                 }}
+                // @ts-expect-error fix this later
                 columns={DATATABLE_COLUMNS}
                 defaultSortOrder={{ name: 'at', direction: 'desc' }}
                 getRowDataCallback={fn => (getRowData = fn)}
