+++ src/pages/katalog-saprodi.tsx
@@ -48,6 +48,7 @@ export default function KatalogProdukSaprodi() {
 
             <Datatable
                 apiUrl={'/public/produk-saprodi/datatable/' + warehouse}
+                // @ts-expect-error fix this later
                 columns={columns}
                 defaultSortOrder={{
                     name: 'category_name',
