+++ src/pages/palm-bunches/rea-payments.tsx
@@ -1,5 +1,4 @@
 // types
-import type { MUIDataTableColumn } from 'mui-datatables'
 import type { UUID } from 'crypto'
 import type PalmBunchesReaPaymentDataType from '@/dataTypes/PalmBunchesReaPayment'
 // vendors
@@ -25,6 +24,7 @@ import toDmy from '@/utils/toDmy'
 import ApiUrlEnum from '../../components/PalmBunchesReaPayment/ApiUrlEnum'
 import formatNumber from '@/utils/formatNumber'
 import numberToCurrency from '@/utils/numberToCurrency'
+import { DataTableColumnObject } from 'mui-datatable-delight'
 
 let getRowData: GetRowDataType<PalmBunchesReaPaymentDataType>
 let mutate: MutateType<PalmBunchesReaPaymentDataType>
@@ -341,7 +341,7 @@ function PalmBunchDeliveryRatesCrudWithUseFormData() {
     )
 }
 
-const DATATABLE_COLUMNS: MUIDataTableColumn[] = [
+const DATATABLE_COLUMNS: DataTableColumnObject[] = [
     {
         name: 'uuid',
         label: 'uuid',
